Refactoring NVLS interfaces
